Knowledge-Resource/L7_evaluation_harness_regression_test/eval/requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
torch>=2.2
tqdm>=4.66
